@@ -1,9 +1,10 @@
+import 'package:back_button_interceptor/back_button_interceptor.dart';
 import 'package:flutter/material.dart';
 import 'package:one_context/src/controllers/one_context.dart';
 
 class OneContextWidget extends StatefulWidget {
-  final Widget child;
-  OneContextWidget({Key key, this.child}) : super(key: key);
+  final Widget? child;
+  OneContextWidget({Key? key, this.child}) : super(key: key);
   _OneContextWidgetState createState() => _OneContextWidgetState();
 }
 
@@ -16,6 +17,21 @@ class _OneContextWidgetState extends State<OneContextWidget> {
         showSnackBar: _showSnackBar,
         showModalBottomSheet: _showModalBottomSheet,
         showBottomSheet: _showBottomSheet);
+    BackButtonInterceptor.add(myInterceptor);
+  }
+
+  @override
+  void dispose() {
+    BackButtonInterceptor.remove(myInterceptor);
+    super.dispose();
+  }
+
+  bool myInterceptor(bool stopDefaultButtonEvent, RouteInfo info) {
+    if (OneContext().hasDialogVisible) {
+      OneContext().popDialog();
+      return true;
+    }
+    return false;
   }
 
   @override
@@ -25,59 +41,59 @@ class _OneContextWidgetState extends State<OneContextWidget> {
       body: Builder(
         builder: (innerContext) {
           OneContext().context = innerContext;
-          return widget.child;
+          return widget.child!;
         },
       ),
     );
   }
 
-  Future<T> _showDialog<T>(
-          {bool barrierDismissible = true,
-          Widget Function(BuildContext) builder,
+  Future<T?> _showDialog<T>(
+          {bool? barrierDismissible = true,
+          required Widget Function(BuildContext) builder,
           bool useRootNavigator = true}) =>
-      showDialog<T>(
+      showDialog<T?>(
         context: context,
         builder: (context) => builder(context),
-        barrierDismissible: barrierDismissible,
+        barrierDismissible: barrierDismissible!,
         useRootNavigator: useRootNavigator,
       );
 
   ScaffoldFeatureController<SnackBar, SnackBarClosedReason> _showSnackBar(
-          SnackBar Function(BuildContext) builder) =>
-      Scaffold.of(OneContext().context)
+          SnackBar Function(BuildContext?) builder) =>
+      ScaffoldMessenger.of(OneContext().context!)
           .showSnackBar(builder(OneContext().context));
 
-  Future<T> _showModalBottomSheet<T>(
-      {Widget Function(BuildContext) builder,
-      Color backgroundColor,
-      double elevation,
-      ShapeBorder shape,
-      Clip clipBehavior,
-      bool isScrollControlled = false,
-      bool useRootNavigator = false,
-      bool isDismissible = true}) {
+  Future<T?> _showModalBottomSheet<T>(
+      {required Widget Function(BuildContext) builder,
+      Color? backgroundColor,
+      double? elevation,
+      ShapeBorder? shape,
+      Clip? clipBehavior,
+      bool? isScrollControlled = false,
+      bool? useRootNavigator = false,
+      bool? isDismissible = true}) {
     return showModalBottomSheet<T>(
       context: context,
       builder: builder,
       backgroundColor: backgroundColor,
       clipBehavior: clipBehavior,
       elevation: elevation,
-      isDismissible: isDismissible,
-      isScrollControlled: isScrollControlled,
+      isDismissible: isDismissible!,
+      isScrollControlled: isScrollControlled!,
       shape: shape,
-      useRootNavigator: useRootNavigator,
+      useRootNavigator: useRootNavigator!,
     );
   }
 
   PersistentBottomSheetController<T> _showBottomSheet<T>(
-      {@required Widget Function(BuildContext) builder,
-      Color backgroundColor,
-      double elevation,
-      ShapeBorder shape,
-      Clip clipBehavior}) {
+      {Widget Function(BuildContext)? builder,
+      Color? backgroundColor,
+      double? elevation,
+      ShapeBorder? shape,
+      Clip? clipBehavior}) {
     return showBottomSheet<T>(
-        context: OneContext().context,
-        builder: builder,
+        context: OneContext().context!,
+        builder: builder!,
         backgroundColor: backgroundColor,
         elevation: elevation,
         shape: shape,
