@@ -8,42 +8,46 @@ import 'package:flutter/material.dart';
 import 'package:one_context/src/controllers/one_theme_controller.dart';
 
 class OneContext with NavigatorController, OverlayController, DialogController {
-  static BuildContext _context;
+  static BuildContext? _context;
 
   /// The almost top root context of the app,
   /// use it carefully or don't use it directly!
-  BuildContext get context {
+  BuildContext? get context {
     assert(_context != null, NO_CONTEXT_ERROR);
     return _context;
   }
 
   static bool get hasContext => _context != null;
-  set context(BuildContext newContext) => _context = newContext;
+  set context(BuildContext? newContext) => _context = newContext;
 
   /// If you need reactive changes, do not use OneContext().mediaQuery
   /// Use `MediaQuery.of(context)` instead.
-  MediaQueryData get mediaQuery => MediaQuery.of(context);
+  MediaQueryData get mediaQuery => MediaQuery.of(context!);
 
   /// If you need reactive changes, do not use OneContext().theme
   /// Use `Theme.of(context)` instead.
-  ThemeData get theme => Theme.of(context);
+  ThemeData get theme => Theme.of(context!);
 
   /// If you need reactive changes, do not use OneContext().textTheme
   /// Use `Theme.of(context).textTheme` instead.
   TextTheme get textTheme => theme.textTheme;
-  FocusScopeNode get focusScope => FocusScope.of(context);
+  FocusScopeNode get focusScope => FocusScope.of(context!);
 
   /// Locale
-  Locale get locale => Localizations.localeOf(context);
+  Locale get locale => Localizations.localeOf(context!);
 
   // ThemeMode and ThemeData
   ThemeMode get themeMode => oneTheme.themeMode;
-  ThemeData get themeData => oneTheme.themeData;
-  ThemeData get darkThemeData => oneTheme.darkThemeData;
+  ThemeData? get themeData => oneTheme.themeData;
+  ThemeData? get darkThemeData => oneTheme.darkThemeData;
 
   // Notifiers
-  OneNotificationController oneNotifier;
-  OneThemeController oneTheme;
+  late OneNotificationController oneNotifier;
+  late OneThemeController oneTheme;
+
+  HeroController heroController = HeroController(
+      createRectTween: (begin, end) =>
+          MaterialRectCenterArcTween(begin: begin, end: end));
 
   OneContext._private() {
     oneNotifier = OneNotificationController();
@@ -55,30 +59,30 @@ class OneContext with NavigatorController, OverlayController, DialogController {
 
   /// Register all necessary callbacks from main widget, automatically
   void registerDialogCallback({
-    Future<T> Function<T>(
-            {bool barrierDismissible,
-            Widget Function(BuildContext) builder,
-            bool useRootNavigator})
+    Future<T?> Function<T>(
+            {bool? barrierDismissible,
+            required Widget Function(BuildContext) builder,
+            bool useRootNavigator})?
         showDialog,
-    Future<T> Function<T>(
-            {Widget Function(BuildContext) builder,
-            Color backgroundColor,
-            double elevation,
-            ShapeBorder shape,
-            Clip clipBehavior,
-            bool isScrollControlled,
-            bool useRootNavigator,
-            bool isDismissible})
+    Future<T?> Function<T>(
+            {required Widget Function(BuildContext) builder,
+            Color? backgroundColor,
+            double? elevation,
+            ShapeBorder? shape,
+            Clip? clipBehavior,
+            bool? isScrollControlled,
+            bool? useRootNavigator,
+            bool? isDismissible})?
         showModalBottomSheet,
     ScaffoldFeatureController<SnackBar, SnackBarClosedReason> Function(
-            SnackBar Function(BuildContext) builder)
+            SnackBar Function(BuildContext?) builder)?
         showSnackBar,
     PersistentBottomSheetController<T> Function<T>(
-            {Widget Function(BuildContext) builder,
-            Color backgroundColor,
-            double elevation,
-            ShapeBorder shape,
-            Clip clipBehavior})
+            {Widget Function(BuildContext)? builder,
+            Color? backgroundColor,
+            double? elevation,
+            ShapeBorder? shape,
+            Clip? clipBehavior})?
         showBottomSheet,
   }) {
     registerCallback(
@@ -98,12 +102,12 @@ class OneContext with NavigatorController, OverlayController, DialogController {
   ///       builder: OneContext().builder,
   ///      ...
   /// ```
-  Widget builder(BuildContext context, Widget widget,
-          {Key key,
-          MediaQueryData mediaQueryData,
-          String initialRoute,
-          Route<dynamic> Function(RouteSettings) onGenerateRoute,
-          Route<dynamic> Function(RouteSettings) onUnknownRoute,
+  Widget builder(BuildContext context, Widget? widget,
+          {Key? key,
+          MediaQueryData? mediaQueryData,
+          String? initialRoute,
+          Route<dynamic> Function(RouteSettings)? onGenerateRoute,
+          Route<dynamic> Function(RouteSettings)? onUnknownRoute,
           List<NavigatorObserver> observers = const <NavigatorObserver>[]}) =>
       ParentContextWidget(
         child: widget,
@@ -116,15 +120,15 @@ class OneContext with NavigatorController, OverlayController, DialogController {
 }
 
 class ParentContextWidget extends StatelessWidget {
-  final MediaQueryData mediaQueryData;
-  final String initialRoute;
-  final Route<dynamic> Function(RouteSettings) onGenerateRoute;
-  final Route<dynamic> Function(RouteSettings) onUnknownRoute;
+  final MediaQueryData? mediaQueryData;
+  final String? initialRoute;
+  final Route<dynamic> Function(RouteSettings)? onGenerateRoute;
+  final Route<dynamic> Function(RouteSettings)? onUnknownRoute;
   final List<NavigatorObserver> observers;
-  final Widget child;
+  final Widget? child;
 
   const ParentContextWidget(
-      {Key key,
+      {Key? key,
       this.child,
       this.mediaQueryData,
       this.initialRoute,
@@ -167,6 +171,6 @@ const String NO_CONTEXT_ERROR = """
   e.g. 
 
   ```dart
-    if (OneContext.hasContext) {OneContext ().showDialog (...);}
+    if (OneContext.hasContext) {OneContext().showDialog (...);}
   ```
 """;
